Move LobbyMenu.java to be grouped with the other menu bar classes
